Rename feed-sort-options.js to feed-options.js
