+++ src/redux/middlewares.js
@@ -6,7 +6,7 @@ import { getPost } from '../services/api';
 import { setToken, persistUser } from '../services/auth';
 import { Connection, scrollCompensator } from '../services/realtime';
 import { userParser, delay } from '../utils';
-import * as FeedSortOptions from '../utils/feed-sort-options';
+import * as FeedOptions from '../utils/feed-options';
 
 import {
   colorSchemeStorageKey,
@@ -29,11 +29,11 @@ export const feedSortMiddleware = (store) => (next) => (action) => {
     const { sort: currentFeedSort, currentFeed } = state.feedSort;
     const { homeFeedSort } = state.user.frontendPreferences;
     if (currentFeed === getFeedName(action)) {
-      action.payload.sortChronologically = currentFeedSort === FeedSortOptions.CHRONOLOGIC;
+      action.payload.sortChronologically = currentFeedSort === FeedOptions.CHRONOLOGIC;
     } else {
       //use home feed setting if we get back to home feed
       //this change isn't yet in reducer, and we don't get it there before real feed request fires
-      action.payload.sortChronologically = action.type === ActionTypes.HOME && homeFeedSort === FeedSortOptions.CHRONOLOGIC;
+      action.payload.sortChronologically = action.type === ActionTypes.HOME && homeFeedSort === FeedOptions.CHRONOLOGIC;
     }
   }
   if (action.type === ActionTypes.TOGGLE_FEED_SORT) {
@@ -423,7 +423,7 @@ const dispatchWithPost = async (store, postId, action, filter = () => true, maxD
   let state = store.getState();
   const shouldBump = isFirstPage(state)
     && !isMemories(state)
-    && state.feedSort.sort === FeedSortOptions.ACTIVITY;
+    && state.feedSort.sort === FeedOptions.ACTIVITY;
 
   if (isPostLoaded(state, postId)) {
     return store.dispatch({ ...action, shouldBump });
@@ -476,7 +476,7 @@ const bindHandlers = (store) => ({
     let insertBefore = null;
     if (shouldBump) {
       insertBefore = state.feedViewState.visibleEntries[0] || null;
-      if (state.feedSort.sort === FeedSortOptions.CHRONOLOGIC) {
+      if (state.feedSort.sort === FeedOptions.CHRONOLOGIC) {
         for (const postId of state.feedViewState.visibleEntries) {
           if (data.posts.createdAt >= state.posts[postId].createdAt) {
             insertBefore = postId;
